fixed merge conflicts in Index
