+++ CLI/.env
@@ -1,0 +1,3 @@
+GITHUB_TOKEN=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
+LOG_LEVEL=2
+LOG_FILE=/home/shay/a/lin1285/ece461/ECE461Phase2/CLI/cache/path.txt
