+++ CLI/run
@@ -1,5 +1,5 @@
 #!/bin/bash
-
+#export $(cat .env | xargs)
 #log environment variables $LOG_PATH and $LOG_LEVEL
 loglevel="${LOG_LEVEL:-1}"
 logfullpath="${LOG_FILE:-/home/shay/a/lin1285/ECE461SoftwareEngineeringProject/cache/log2.txt}"
